Massive commit! Near rewrite status. See full log
This turned out to big commit. The only target of this huge change was to
split the modules off into their own file if possible. This, I think, was
accomplished.
A helper function util::join was added.
!value was also added to get the value of a variable
A few things were moved into config::, along with a few things into global::