Continuing in the long strain of commits, see full log
This sets up the reading and writing of serialized Function objects. Included
in this is the code to properly read and write the markov chain database. It
is probably larger than we would hope, but it'll have to do for now :)

The other change was to move IgnoreFunction into modules/core so that it can
be read/written properly