Another huge commit, see full log
The purpopse of this changeset is to facilitate the script module(s).

The help function was moved into modules/core
Most of the variables have been moved into config::
siMap was moved out of FunctionArguments and into global::
Line parsing was made it's own function in global:: along with logging
the !on function simulates a line when a regex matches.
Some other fiddling had to be done