Major restructuring of markov model, see full log
This change is a move towards lowering redundancy as much as possible. Before,
for any given text of length "n", there was a copy of each word individually
for the zero order data. Then two more copies of each word that appeared in
sequence. Finally, another copy for every three word sequence.

Now, the words are stored once and converted to/from integers. Currently there
is still a lot of data duplication. Dictionary will be turned into a static
minimal discrete acylclic finite-state automaton to use for a perfect hashing
scheme for the previously known words (rebuilt everytime on load). There will
still be the current Dictionary internals to keep track of new words
encountered during the lifetime of the program.

MarkovModel is no longer a simple string -> int map, but now a templated
recursive class. Most of the old functionality is back. However, count and
correct are both currently broken.