First draft of new module parsing system; see full log
bin/makemods greps through source files in modules/ to find applicable modules
and functions. Currently it works with markov, but will most likely need
changing when the other modules are added and mistakes are seen (like module
name detection)

The Makefile attempts to run makemods whenever necessary, but may fail. If in
doubt, run it yourself.

A lot of global has been commented so that it compiles. This code will be
replaced entirely soon.