Restructure for single-binary output, add Tupfile
These changes are the start of simplifying dependency tracking and the build
system. We remove some redundant build rules by having only a single binary
that accepts switches instead of four separate binaries, which also enables us
to add a Tupfile which properly builds the project.

More changes to come...