Basic support for generic backtrace support on errors
This includes the err::ex class and the except() helper macro. These throw an
exception which retains backtrace information, and provides a programmatic
interface to the catcher of where the error happened.

A first draft of prettytrace is included, which maps from the backtrace()
returned binary addresses into files, line numbers, and functions. It is
rather slow and calls out to addr2line too much, but it works.