Fix spacing in args variable; restore function arguments on ;
This fixes two longstanding bugs. The first, the args variable no longer has
an extra leading space.

The second, during function execution it was often possible for the $ and args
variables to get clobbered by internal function calls. This may still be
possible depending on the way subfunctions are chained, but at least
semicolons give sequence points where we know the args are restored.