Produce much better StackTrace results in pbrane
This change scratches a number of itches. We now have a max recursion depth.
Simple things like parenthesis do not count against it, even though we have a
stackframe internally for them. Frames that are complete are removed from the
stack trace -- instead of producing a stack graph. Function calls are
abbreviated once we know the function name.